extract mega-moe to layers/moe/mega_moe.py
