+++ python/sglang/srt/layers/quantization/fp8.py
@@ -1166,7 +1166,7 @@ class Fp8MoEMethod(FusedMoEMethodBase):
                 layer.w2_weight.data = layer.w2_weight.data.view(torch.int8)
 
                 if envs.SGLANG_OPT_USE_DEEPGEMM_MEGA_MOE.get():
-                    from sglang.srt.models.deepseek_v4 import (
+                    from sglang.srt.layers.moe.mega_moe import (
                         build_mega_moe_experts_weights,
                     )
