+++ python/sglang/srt/models/deepseek_v4.py
@@ -1698,71 +1698,6 @@ def _dequant_fp8(weight: torch.Tensor, scale: torch.Tensor) -> torch.Tensor:
     return result.to(torch.bfloat16)
 
 
-def build_mega_moe_experts_weights(experts) -> None:
-    from deep_gemm import (
-        transform_sf_into_required_layout,
-        transform_weights_for_mega_moe,
-    )
-    from deep_gemm.mega import _interleave_l1_weights, _transpose_sf_for_utccp
-
-    if getattr(experts, "_mega_moe_weights_built", False):
-        return
-
-    w13 = experts.w13_weight.data
-    w13_sf_fp32 = experts.w13_weight_scale_inv.data
-    w2 = experts.w2_weight.data
-    w2_sf_fp32 = experts.w2_weight_scale_inv.data
-
-    num_groups, n1, half_k1 = w13.shape
-    k1 = half_k1 * 2
-    _, n2, half_k2 = w2.shape
-    k2 = half_k2 * 2
-
-    w13_sf = transform_sf_into_required_layout(
-        w13_sf_fp32,
-        mn=n1,
-        k=k1,
-        recipe=(1, 32),
-        num_groups=num_groups,
-        disable_ue8m0_cast=False,
-    )
-    w2_sf = transform_sf_into_required_layout(
-        w2_sf_fp32,
-        mn=n2,
-        k=k2,
-        recipe=(1, 32),
-        num_groups=num_groups,
-        disable_ue8m0_cast=False,
-    )
-
-    if envs.SGLANG_OPT_FIX_MEGA_MOE_MEMORY.get():
-        # Build the interleaved L1 weight + scale once; share the weight buffer
-        # between `w13_weight.data` (normal deep-ep path) and `mega_l1_weights[0]`
-        # (mega moe path). Mega moe additionally needs a UTCCP-transposed scale;
-        # the deep-ep path consumes the non-transposed interleaved scale and a
-        # swizzle-aware activation kernel. L2 weight is untouched by the mega
-        # transform, so the existing `w2_weight.data` is shared directly.
-        w13_interleaved, w13_sf_interleaved = _interleave_l1_weights((w13, w13_sf))
-        w13_sf_utccp = _transpose_sf_for_utccp(w13_sf_interleaved)
-        w2_sf_utccp = _transpose_sf_for_utccp(w2_sf)
-
-        experts.w13_weight.data = w13_interleaved
-        experts.w13_weight_scale_inv.data = w13_sf_interleaved
-        experts.w2_weight_scale_inv.data = w2_sf
-        experts.w13_weight_scale_inv.format_ue8m0 = True
-        experts.w2_weight_scale_inv.format_ue8m0 = True
-
-        experts.mega_l1_weights = (experts.w13_weight.data, w13_sf_utccp)
-        experts.mega_l2_weights = (experts.w2_weight.data, w2_sf_utccp)
-    else:
-        l1_pair, l2_pair = transform_weights_for_mega_moe((w13, w13_sf), (w2, w2_sf))
-
-        experts.mega_l1_weights = l1_pair
-        experts.mega_l2_weights = l2_pair
-
-    experts._mega_moe_weights_built = True
-
-
 def _dequant_fp8_wo_a(
     weights: Iterable[Tuple[str, torch.Tensor]],
 ) -> Iterable[Tuple[str, torch.Tensor]]:
